+++ examples/cuda/convolution.py
@@ -10,12 +10,14 @@ def tune():
 
     #setup tunable parameters
     tune_params = OrderedDict()
-    tune_params["filter_height"] = [i for i in range(3,35,2)]
-    tune_params["filter_width"] = [i for i in range(3,35,2)]
+    #tune_params["filter_height"] = [i for i in range(3,35,2)]
+    tune_params["filter_height"] = [17]
+    #tune_params["filter_width"] = [i for i in range(3,35,2)]
+    tune_params["filter_width"] = [17]
     tune_params["block_size_x"] = [16*i for i in range(1,9)]
     tune_params["block_size_y"] = [2**i for i in range(6)]
-    tune_params["tile_size_x"] = [2**i for i in range(3)]
-    tune_params["tile_size_y"] = [2**i for i in range(3)]
+    tune_params["tile_size_x"] = [i for i in range(1,9)]
+    tune_params["tile_size_y"] = [i for i in range(1,9)]
 
     tune_params["use_padding"] = [0,1]  #toggle the insertion of padding in shared memory
     tune_params["read_only"] = [0,1]    #toggle using the read-only cache
@@ -40,15 +42,23 @@ def tune():
     grid_div_y = ["block_size_y", "tile_size_y"]
 
     #start tuning
-    return kernel_tuner.tune_kernel("convolution_kernel", kernel_string,
+    import time
+
+    start = time.time()
+
+    results, env = kernel_tuner.tune_kernel("convolution_kernel", kernel_string,
         problem_size, args, tune_params,
         grid_div_y=grid_div_y, grid_div_x=grid_div_x, cmem_args=cmem_args,
         verbose=True, restrictions=restrict)
 
+    end = time.time()
+    env['execution_time'] = end-start
+
+    return results, env
 
 if __name__ == "__main__":
     results = tune()
     import json
-    with open("convolution.json", 'w') as fp:
+    with open("convolution_17x17.json", 'w') as fp:
         json.dump(results, fp)